backend/requirements.txt
fastapi
uvicorn
gitpython
radon
cloc 